# docs/openapi.yaml
openapi: 3.1.0
info:
  title: Outpost VPN API
  description: |
    Open-source WireGuard VPN & Zero-Trust Network Access platform.
    All features available under Apache 2.0 — no enterprise paywalls.
  version: 1.0.0
  license:
    name: Apache 2.0
    url: https://www.apache.org/licenses/LICENSE-2.0
  contact:
    name: Romashqua Labs
    url: https://github.com/romashqua/outpost
servers:
  - url: http://localhost:8080
    description: Local development
  - url: https://vpn.example.com
    description: Production
...
paths:
...
  /api/v1/smart-routes:
    get:
      operationId: listSmartRoutes
      summary: List smart routes
      tags: [SmartRoutes]
      responses:
        "200":
          description: List of smart routes
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/SmartRoute"
        "500":
          description: Internal error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
    post:
      operationId: createSmartRoute
      summary: Create a smart route
      tags: [SmartRoutes]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateSmartRouteRequest"
      responses:
        "201":
          description: Smart route created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SmartRoute"
        "400":
          description: Missing name
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "409":
          description: Duplicate name
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "500":
          description: Internal error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /api/v1/smart-routes/{id}:
    get:
      operationId: getSmartRoute
      summary: Get a smart route with entries
      tags: [SmartRoutes]
      parameters:
        - $ref: "#/components/parameters/ResourceID"
      responses:
        "200":
          description: Smart route with entries
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SmartRoute"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
    put:
      operationId: updateSmartRoute
      summary: Update a smart route
      tags: [SmartRoutes]
      parameters:
        - $ref: "#/components/parameters/ResourceID"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateSmartRouteRequest"
      responses:
        "200":
          description: Smart route updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SmartRoute"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "409":
          description: Duplicate name
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
    delete:
      operationId: deleteSmartRoute
      summary: Delete a smart route
      tags: [SmartRoutes]
      parameters:
        - $ref: "#/components/parameters/ResourceID"
      responses:
        "204":
          description: Deleted
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /api/v1/smart-routes/{id}/entries:
    post:
      operationId: addSmartRouteEntry
      summary: Add an entry to a smart route
      tags: [SmartRoutes]
      parameters:
        - $ref: "#/components/parameters/ResourceID"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateSmartRouteEntryRequest"
      responses:
        "201":
          description: Entry added
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SmartRouteEntry"
        "400":
          description: Invalid fields or missing proxy_id
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "409":
          description: Duplicate entry
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "500":
          description: Internal error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /api/v1/smart-routes/{id}/entries/{entryId}:
    delete:
      operationId: deleteSmartRouteEntry
      summary: Remove an entry from a smart route
      tags: [SmartRoutes]
      parameters:
        - $ref: "#/components/parameters/ResourceID"
        - name: entryId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "204":
          description: Entry deleted
        "404":
          description: Entry not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /api/v1/smart-routes/proxy-servers:
    get:
      operationId: listProxyServers
      summary: List proxy servers
      tags: [SmartRoutes]
      responses:
        "200":
          description: List of proxy servers
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ProxyServer"
        "500":
          description: Internal error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
    post:
      operationId: createProxyServer
      summary: Create a proxy server
      tags: [SmartRoutes]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateProxyServerRequest"
      responses:
        "201":
          description: Proxy server created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProxyServer"
        "400":
          description: Missing or invalid fields
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "409":
          description: Duplicate proxy server
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "500":
          description: Internal error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /api/v1/smart-routes/proxy-servers/{id}:
    delete:
      operationId: deleteProxyServer
      summary: Delete a proxy server
      tags: [SmartRoutes]
      parameters:
        - $ref: "#/components/parameters/ResourceID"
      responses:
        "204":
          description: Deleted
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "409":
          description: Referenced by route entries
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"

  # ── Webhooks ──────────────────────────────────────────────────────────
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT token obtained via POST /api/v1/auth/login
  parameters:
    ResourceID:
      name: id
      in: path
      required: true
      schema:
        type: string
        format: uuid
...
  schemas:
...
    Error:
      type: object
      required: [error]
      properties:
        error:
          type: string
          example: something went wrong

    # ── Auth ────────────────────────────────────────────────────────
...
    SmartRoute:
      type: object
      required: [id, name, is_active, created_at, updated_at]
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        description:
          type: string
          nullable: true
        is_active:
          type: boolean
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
        entries:
          type: array
          items:
            $ref: "#/components/schemas/SmartRouteEntry"
    SmartRouteEntry:
      type: object
      required: [id, smart_route_id, entry_type, value, action, priority, created_at]
      properties:
        id:
          type: string
          format: uuid
        smart_route_id:
          type: string
          format: uuid
        entry_type:
          type: string
          enum: [domain, cidr, domain_suffix]
        value:
          type: string
          example: "*.google.com"
        action:
          type: string
          enum: [proxy, direct, block]
        proxy_id:
          type: string
          format: uuid
          nullable: true
        proxy_name:
          type: string
          nullable: true
        priority:
          type: integer
          example: 100
        created_at:
          type: string
          format: date-time
    CreateSmartRouteRequest:
      type: object
      required: [name]
      properties:
        name:
          type: string
          minLength: 1
        description:
          type: string
    UpdateSmartRouteRequest:
      type: object
      properties:
        name:
          type: string
        description:
          type: string
        is_active:
          type: boolean
    CreateSmartRouteEntryRequest:
      type: object
      required: [entry_type, value, action]
      properties:
        entry_type:
          type: string
          enum: [domain, cidr, domain_suffix]
        value:
          type: string
        action:
          type: string
          enum: [proxy, direct, block]
        proxy_id:
          type: string
          format: uuid
          description: Required when action is 'proxy'
        priority:
          type: integer
          default: 100
    ProxyServer:
      type: object
      required: [id, name, type, address, port, is_active, created_at, updated_at]
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        type:
          type: string
          enum: [socks5, http, shadowsocks, vless]
        address:
          type: string
        port:
          type: integer
        username:
          type: string
          nullable: true
        password:
          type: string
          nullable: true
        extra_config:
          type: string
          nullable: true
          description: JSON-encoded extra configuration
        is_active:
          type: boolean
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
    CreateProxyServerRequest:
      type: object
      required: [name, type, address, port]
      properties:
        name:
          type: string
          minLength: 1
        type:
          type: string
          enum: [socks5, http, shadowsocks, vless]
        address:
          type: string
        port:
          type: integer
          minimum: 1
          maximum: 65535
        username:
          type: string
        password:
          type: string
        extra_config:
          type: string
          description: JSON string

    # ── Webhook ───────────────────────────────────────────────────────
...
security:
  - BearerAuth: []